server/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
openenv-core>=0.2.1
websockets>=11.0
requests>=2.25.0
python-multipart>=0.0.6